requirements.txt
requests
pyyaml
configobj
click
jinja2
lxml